chore(env-utils): move ua-parser-js dependency to suite
