+++ packages/env-utils/package.json
@@ -28,9 +28,6 @@
         "build:lib": "yarn g:rimraf ./lib && yarn g:tsc --build tsconfig.lib.json && ../../scripts/publish/replace-imports.sh ./lib cjs",
         "prepublishOnly": "yarn tsx ../../scripts/publish/prepublishNPM.js"
     },
-    "dependencies": {
-        "ua-parser-js": "^2.0.4"
-    },
     "peerDependencies": {
         "expo-constants": "*",
         "expo-localization": "*",
