+++ packages/env-utils/src/index.ts
@@ -10,14 +10,8 @@ export const {
     getUserAgent,
     isAndroid,
     isChromeOs,
-    getBrowserVersion,
-    getBrowserName,
     getCommitHash,
-    getDeviceType,
-    getOsVersion,
-    getCpuArch,
     getSuiteVersion,
-    isFirefox,
     getPlatform,
     getPlatformLanguages,
     getScreenWidth,
@@ -33,8 +27,6 @@ export const {
     isLinux,
     isCodesignBuild,
     getOsName,
-    getOsNameWeb,
-    getOsFamily,
     getJWSPublicKey,
 } = envUtils;
