+++ suite-common/suite-utils/src/index.ts
@@ -10,3 +10,4 @@ export * from './jws';
 export * from './pollingController';
 export * from './triggerWebDownloadFile';
 export * from './translations';
+export * from './userAgent';
